+++ internal/engine/schedule.go
@@ -67,12 +67,20 @@ func (s *Schedule) Schedule() {
 func (s *Schedule) CreateWork() {
 	for {
 		r := <-s.workerCh
+		if err := r.Check(); err != nil {
+			s.Logger.Error("check failed",
+				zap.Error(err),
+			)
+			continue
+		}
+
 		body, err := s.Fetcher.Get(r)
 		if len(body) < 6000 {
 			s.Logger.Error("can't fetch ",
 				zap.Int("length", len(body)),
 				zap.String("url", r.Url),
 			)
+			continue
 		}
 		if err != nil {
 			s.Logger.Error("can't fetch ",